requirements.txt
Flask==3.0.2
functions-framework==3.*
flask-cors==4.*
requests==2.26.0
pymongo==4.3.3
openai==0.27.8
google-cloud-storage==2.5.0
python-telegram-bot==20.3
python-dotenv==1.0.0
python-magic==0.4.27